+++ docs/parser/jvm-object-generation.md
@@ -20,10 +20,13 @@ the compiler and runtime to work with the AST.
 
 # Overall Architecture
 
-The JVM foreign function interface (FFI) named [Java Native Interface](
-https://en.wikipedia.org/wiki/Java_Native_Interface) (JNI) enables the 
+The JVM foreign function interface (FFI) named 
+[Java Native Interface](https://en.wikipedia.org/wiki/Java_Native_Interface) 
+(JNI) enables the 
 Rust parser library to call and be called by the parser library
-implemented on top of Scala. Specifically, the Scala application uses
-JNI to call functions exposed by Rust parser, while the Rust parser
-itself uses the JNI to call Scala AST constructors while consuming
-the input. 
+implemented on top of Scala.
+
+Specifically, in our architecture the JNI is used for:
+- Invoking Rust parser methods from Scala.
+- Invoking Scala AST constructors from Rust.
+
